add nextauth
in order to make "useSession" calls, i need to mark the component
as being a client component. i also need to wrap the app with a
<SessionProvider>, but the problem is that the root layout.tsx
component CAN'T be a client component. the recommended approach is to
create a separate context that can be marked a client component,
and then importing that into the root layout.tsx:

https://beta.nextjs.org/docs/rendering/server-and-client-components#using-context-in-client-components

i also made an attempt to set up nextauth with the new "app" directory
but was having issues since i don't think nextauth is currently
prepared to handle those updates with the `route.ts` approach.